[IMP] abstract common logic and open source DevCO crm geo.
