+++ l10n_co_geo_crm/README.md
@@ -1,0 +1,26 @@
+Colombian Geo CRM
+=================
+
+This module was written to extend the functionality of DevCO's Colombian Basic Accounting (l10n_co) to support colombian departments and municipalities with their corresponding official codes. It also aims to improve usability in adressfields and helps with disambiguation (ej. Armenia, Quindio vs. Armenia, Antioquia). This is the CRM part of the l10n_co_geo base module.
+
+Credits
+=======
+
+Contributors
+------------
+
+* David Arnold <dar@devco.co>
+* Juan Pablo Arias <jpa@devco.co>
+
+Maintainer
+----------
+
+.. image:: https://www.devco.co/logo.png
+   :alt: DevCO Colombia SAS
+   :target: http://devco.co
+
+This module is maintained by DevCO Colombia SAS.
+
+DevCO Colombia is a Colombian Odoo Specialist whose mission is to provide the most comprehensive colombian localisation and a broad range of next generation hosting and implementation services.
+
+For further information, please visit <a href="http://devco.co">http://devco.co</a> .
