+++ l10n_co_geo_crm/__openerp__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+{
+    'version': '1.0',
+    'name': 'Colombian Geo CRM',
+    'summary': 'CRM Adress magic upon cities, departments and countries.',
+    'category': 'Localization',
+
+    'author': 'DevCO',
+    'website': 'https://www.felicity.com.co',
+    'license': 'LGPL-3',
+
+    'depends': ['l10n_co_geo', 'crm'],
+    'data': ['views/crm_lead_view.xml'],
+    'auto_install': True,
+}
