+++ l10n_co_geo_crm/models/__init__.py
@@ -2,9 +2,4 @@
 # Copyright 2016 David Arnold, DevCO Colombia
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
-from . import (
-    partner_city_abstract,
-    res_country,
-    res_partner,
-    res_company
-)
+from . import crm_lead
